+++ NEWS.md
@@ -15,6 +15,7 @@
 * Improve header checks to be more RFC compliant (#78)
 * Use `httr`s `RETRY` function for retries that handles handles better (#67)
 * Fix `http2 framing layer error` that sometimes happens in misconfigured http2 servers (#87)
+* Update minimum dependcy to R 3.3.0 to support Shiny function bugs. 
  
 # v0.5.1